% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{app}
\alias{app}
\title{TIFF App server function.}
\usage{
app(input, output, session, fm, settings, geneSignatures)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{fm}{future manager object}

\item{settings}{cliff settings}

\item{geneSignatures}{result of the \code{getAvailableGeneSignatures}}
}
\value{
shiny app server
}
\description{
TIFF App server function.
}
\details{
It should not be called directly by the user.
Please use \code{TIFF::run()}.
}
