% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app.R
\name{appUI}
\alias{appUI}
\title{TIFF App UI}
\usage{
appUI(settings, proliferationTests, hallmarkGeneSets)
}
\arguments{
\item{settings}{settings object}

\item{proliferationTests}{proliferation tests config extracted from the settings
object by \code{getAvailableProliferationTests(getDrugScoreConfig(settings))}}

\item{hallmarkGeneSets}{list of available genesets.
E.g. a result of \code{getAvailableHallmarkGeneSets()}}
}
\description{
TIFF App UI
}
