% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{compareClones}
\alias{compareClones}
\title{Compare Clones}
\usage{
compareClones(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame
}
\description{
Compare Clones
}
\examples{

setDbOptions(getSettings())

ca <- exampleClassAssigment(high = 0.16)
compareClones(ca)

}
