% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{compareImmuneCells}
\alias{compareImmuneCells}
\title{Compare Immune Cells}
\usage{
compareImmuneCells(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame with columns `cell type`, method, log2FC, p.value, adj.P.Value
}
\description{
Compare Immune Cells
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.16)
head(compareImmuneCells(ca))

}
