% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{compareMetabolics}
\alias{compareMetabolics}
\title{Compare Metabolics}
\usage{
compareMetabolics(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame with columns `metabolic pathway`, method, diff, p.value, adj.P.Value
}
\description{
Compare Metabolics
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.16)
head(compareMetabolics(ca))

}
