% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{compareSignaling}
\alias{compareSignaling}
\title{Compare Signaling}
\usage{
compareSignaling(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame with columns pathway, class1.n_active, class1.n_total,
class1.percentage, class2.n_active, class2.n_total, class2.percentage, higher, p.value
}
\description{
Compare Signaling
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.16)
head(compareSignaling(ca))

}
