% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{differentialGeneExpression_LimmaVoom}
\alias{differentialGeneExpression_LimmaVoom}
\title{Differential gene expression}
\usage{
differentialGeneExpression_LimmaVoom(tissueClasses, anno, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{anno}{data.frame, gene annotation}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame
}
\description{
This function uses limma package to determine the differentially expressed
genes or transcripts across 2 classes.
}
\details{
If you want to report progress, please provide a value for `p` argument. See 
`ProcessProgress` for more details.

TODO: Add tpmHET. It's not supported yet.

As a result, this function returns a data.frame
contains the statistics for genes or transcripts. 
`correctTpms` is a column of determining if gene or transcript ID is  
highly expressed, i.e. tpm > `tpmHET`
(`tpmHET` - tpm High Expression Threshold, default 10, it can be set in the
`differentialGeneExpression_LimmaVoom` function or by using `cliff.tpmHET` 
option) for at least 1 tissue.


Columns available in `df`: ensg or enst, `anno` columns, P.Value, adj.p.val, 
logFC, auc.
}
\examples{
\dontrun{
setDbOptions(getSettings())
tissueClasses <- classAssignment(sensitive = c("a", "b", "c"), resistant = c("d", "e", "f"))
gAnno <- getGeneAnno("human")

gRes <- differentialGeneExpression_LimmaVoom(tissueClasses, gAnno, p = TRUE)
}
}
