% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{differentialHallmarkSets}
\alias{differentialHallmarkSets}
\title{Differential Hallmark sets}
\usage{
differentialHallmarkSets(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
tibble with columns: gene_set, higher, logFC, P.Value, adj.p.val
}
\description{
Compares Hallmark sets data for 2 tissue classes.
}
\examples{
\dontrun{
setDbOptions(getSettings())
tissueClasses <- classAssignment(sensitive = c("a", "b", "c"), resistant = c("d", "e", "f"))

differentialHallmarkSets(tissueClasses, p = TRUE)
}
}
