% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{differentialProteinExpression}
\alias{differentialProteinExpression}
\title{Differential protein expression}
\usage{
differentialProteinExpression(tissueClasses, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
tibble with columns: antibody, higher, log2FC, P.Value, adj.p.val
}
\description{
Compares protein expression data for 2 tissue classes.
}
\examples{
setDbOptions(getSettings())
tissueClasses <- exampleClassAssigment(high = 1)

differentialProteinExpression(tissueClasses, p = TRUE)

}
