% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{exampleClassAssigment}
\alias{exampleClassAssigment}
\title{Example Tissue Class Assigment Object}
\usage{
exampleClassAssigment(
  low = 0.15,
  high = 10,
  tt = c("cholangiocarcinoma", "colon carcinoma", "melanoma", "ovarian carcinoma",
    "renal carcinoma", "rhabdomyosarcoma", "gastric carcinoma")
)
}
\arguments{
\item{low}{lower threshold for mutational_fraction_in_percent}

\item{high}{upper threshold for mutational_fraction_in_percent}

\item{tt}{tumor types}
}
\value{
ClassAssigment object
}
\description{
Example Tissue Class Assigment Object
}
\examples{

setDbOptions(getSettings())
exampleClassAssigment()

}
