% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateClonesPlot}
\alias{generateClonesPlot}
\title{Generate Clones Plot}
\usage{
generateClonesPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{result of getTissueDataClones with score column}

\item{plotType}{plot type: point, roc, violin, box, coverage or all.}

\item{title}{plot title}

\item{ca}{classAssigment object (in most cases it can be
inherited from df)}
}
\value{
ggplot object
}
\description{
Generate Clones Plot
}
\examples{

setDbOptions(getSettings())

ca <- exampleClassAssigment(high = 0.16)
df <- getTissueDataClones(ca, "number_of_clones") \%>\%
 mutate(score = `number_of_clones`)
generateClonesPlot(df, "all")

}
