% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateClonesWaterfallPlot}
\alias{generateClonesWaterfallPlot}
\title{Generate Clones Waterfall Plot}
\usage{
generateClonesWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame created with \code{getWaterfallDataClones}}

\item{fill}{column from data used to fill value}
}
\value{
ggplot2 object
}
\description{
Generate Clones Waterfall Plot
}
\examples{

setDbOptions(getSettings())

prefilter <- makePrefilter()
df <- getWaterfallDataClones("number_of_clones", prefilter)
generateClonesWaterfallPlot(df)

}
