% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateCopynumberExpressionPlot}
\alias{generateCopynumberExpressionPlot}
\title{Generate CN vs gene expression plot}
\usage{
generateCopynumberExpressionPlot(dfCN, dfExpr, title = NULL, ca = NULL)
}
\arguments{
\item{dfCN}{data.frame, output of getTissueDataCopyNumberById()}

\item{dfExpr}{data.frame, output of getTissueDataGeneExpressionById()}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{dfCN} param.
However if dfCN is an output of getTissueDataCopyNumberById, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces comparison plot with copy number and gene expression data.
}
\examples{
setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"
dfCN <- getTissueDataCopyNumberById(ensg, ca)
dfExpr <- getTissueDataGeneExpressionById(ensg, ca)
generateCopynumberExpressionPlot(dfCN, dfExpr)

}
