% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateCopynumberPlot}
\alias{generateCopynumberPlot}
\title{Generate copy number plot by type}
\usage{
generateCopynumberPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame, output of getTissueDataCopyNumberById()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataCopyNumberById, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces copy number plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage" and "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"
df <- getTissueDataCopyNumberById(ensg, ca)
symbol <- XIFF::getGeneSymbol(ensg)
generateCopynumberPlot(df, "point", paste(symbol, "point plot"))
generateCopynumberPlot(df, "roc", paste(symbol, "roc plot"))
generateCopynumberPlot(df, "violin", paste(symbol, "violin plot"))
generateCopynumberPlot(df, "box", paste(symbol, "box plot"))
generateCopynumberPlot(df, "coverage", paste(symbol, "data coverage"))

}
