% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateExpressionHeatmap}
\alias{generateExpressionHeatmap}
\title{Generate Expression Heatmap}
\usage{
generateExpressionHeatmap(
  genes,
  sampleClasses,
  sampleAnno,
  sampleLabel = "tissuename",
  annoVariables = NULL,
  heat_scale = "genes",
  flipAxes = FALSE,
  classLabel = NULL
)
}
\arguments{
\item{genes}{character vectors of genes}

\item{sampleClasses}{classAsigment object. In most cases it should be inferred from 
\code{data}. However if it is not possible, you may need to provide
it on your own.}

\item{sampleAnno}{tissue annotation data.frame}

\item{heat_scale}{}

\item{flipAxes}{logical. Should the axes be flipped.}
}
\value{

}
\description{
Generate Expression Heatmap
}
\examples{

setDbOptions(getSettings())

ca <- exampleClassAssigment()
gene_set <- getGSEAdata("human", gene_set = "HALLMARK_P53_PATHWAY")
cl_anno <- getTissueAnnotation()

annoVariables <- c("tumortype")

plot.new()
generateExpressionHeatmap(
 genes = gene_set, 
 sampleClasses = ca, 
 sampleAnno = cl_anno,
 annoVariables = annoVariables
)

plot.new()
generateExpressionHeatmap(
 genes = gene_set, 
 sampleClasses = ca, 
 sampleAnno = cl_anno,
 annoVariables = annoVariables,
 sampleLabel = "tumortype"
)

plot.new()
generateExpressionHeatmap(
 genes = gene_set, 
 sampleClasses = ca, 
 sampleAnno = cl_anno,
 annoVariables = c("tumortype", "organ"),
 sampleLabel = "tumortype"
)

}
