% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateExpressionPlot}
\alias{generateExpressionPlot}
\title{Generate gene expression plot by type}
\usage{
generateExpressionPlot(df, plotType = "point", title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame, output of getTissueDataGeneExpressionById()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataGeneExpressionById, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces expression plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage" and "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"
df <- getTissueDataGeneExpressionById(ensg, ca)
symbol <- XIFF::getGeneSymbol(ensg)
generateExpressionPlot(df, "point", paste(symbol, "point plot"))
generateExpressionPlot(df, "roc", paste(symbol, "roc plot"))
generateExpressionPlot(df, "violin", paste(symbol, "violin plot"))
generateExpressionPlot(df, "box", paste(symbol, "box plot"))
generateExpressionPlot(df, "coverage", paste(symbol, "data coverage"))
generateExpressionPlot(df, "all")

}
