% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateGeneSetPlot}
\alias{generateGeneSetPlot}
\title{Generate gene set plot by type}
\usage{
generateGeneSetPlot(
  df,
  plotType = "point",
  title = NULL,
  ca = NULL,
  ylabel = "gsva"
)
}
\arguments{
\item{df}{data.frame, output of getTissueDataGeneSetById()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataGeneSetById, then this
parameter does not need to be specified.}

\item{ylabel}{character string, y-axis label}
}
\value{
ggplot object
}
\description{
Produces gene set plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage" and "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
geneSet <- "HALLMARK_APOPTOSIS"
df <- getTissueDataGeneSetById(geneSet, ca)
generateGeneSetPlot(df, "point", paste(geneSet, "point plot"))
generateGeneSetPlot(df, "roc", paste(geneSet, "roc plot"))
generateGeneSetPlot(df, "violin", paste(geneSet, "violin plot"))
generateGeneSetPlot(df, "box", paste(geneSet, "box plot"))
generateGeneSetPlot(df, "coverage", paste(geneSet, "data coverage"))

}
