% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateGeneSignaturePlot}
\alias{generateGeneSignaturePlot}
\title{Generate gene signature plot by type}
\usage{
generateGeneSignaturePlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame, output of getTissueDataGeneSignatureById()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataGeneSignatureById, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces gene signature plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage" and "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
signature <- "MERCK18"
df <- getTissueDataGeneSignatureById(signature, ca)
generateGeneSignaturePlot(df, "point", paste(signature, "point plot"))
generateGeneSignaturePlot(df, "roc", paste(signature, "roc plot"))
generateGeneSignaturePlot(df, "violin", paste(signature, "violin plot"))
generateGeneSignaturePlot(df, "box", paste(signature, "box plot"))
generateGeneSignaturePlot(df, "coverage", paste(signature, "data coverage"))

}
