% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateGeneSignatureWaterfallPlot}
\alias{generateGeneSignatureWaterfallPlot}
\title{Generate waterfall plot for gene signatures}
\usage{
generateGeneSignatureWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame, output of getWaterfallDataGeneSignatures}

\item{fill}{character string, variable to use for bars coloring}
}
\value{
ggplot object
}
\description{
Produces a waterfall plot with gene signature data
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
df <- getWaterfallDataGeneSignatures("MERCK18", prefilter)
generateGeneSignatureWaterfallPlot(df)

}
