% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateHallmarkSetWaterfallPlot}
\alias{generateHallmarkSetWaterfallPlot}
\title{Generate waterfall plot for gene set}
\usage{
generateHallmarkSetWaterfallPlot(data, fill = "tumortype", ylabel = "gsva")
}
\arguments{
\item{data}{data.frame, output of getWaterfallDataHallmarkSet}

\item{fill}{character string, variable to use for bars coloring}

\item{ylabel}{character string, y-axis label}
}
\value{
ggplot object
}
\description{
Produces a waterfall plot with gene set data
}
\examples{
setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
df <- getWaterfallDataHallmarkSet("HALLMARK_APOPTOSIS", prefilter)
generateHallmarkSetWaterfallPlot(df)

}
