% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateImmuneCellsPlot}
\alias{generateImmuneCellsPlot}
\title{Generate Immune Cells Plot}
\usage{
generateImmuneCellsPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame created with \code{\link{getTissueDataImmuneCellsById}}}

\item{plotType}{one of the follwoing "point", "roc", "violin", "box", "coverage"
or "all" to get a list with all plots}

\item{title}{string with the title to be added to the plot}

\item{ca}{classAssigment object (in most cases it will be inherited from df)
however in some cases it needs to be provided explicitly.}
}
\value{
plot or a list of plots if plotType is 'all'
}
\description{
Generate Immune Cells Plot
}
\examples{

setDbOptions(getSettings())
data <- getTissueDataImmuneCellsById("Neurons", exampleClassAssigment(high = 0.16))

type <- c("point", "roc", "violin", "box", "coverage")
generateImmuneCellsPlot(data, "all")

}
