% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateImmuneCellWaterfallPlot}
\alias{generateImmuneCellWaterfallPlot}
\title{Generate Immune Cell Waterfall Plot}
\usage{
generateImmuneCellWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame create with \code{getWaterfallDataImmuneCells}}

\item{fill}{column used to fill the data on the plot}
}
\value{
ggplot2 object
}
\description{
Generate Immune Cell Waterfall Plot
}
\examples{

setDbOptions(getSettings())

df <- getWaterfallDataImmuneCells("Activated dendritic cells", makePrefilter())
generateImmuneCellWaterfallPlot(df)

}
