% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateKaplanMeierPlot}
\alias{generateKaplanMeierPlot}
\title{Generate Kaplan-Meier Plot}
\usage{
generateKaplanMeierPlot(data, plot_p = TRUE, plot_ci = FALSE, plot_risk = TRUE)
}
\arguments{
\item{data}{tiff survival analysis created with \code{\link{survivalAnalysis}}}

\item{plot_p}{if TRUE (default), then the p-value is added to the plot}

\item{plot_ci}{if TRUE (defult FALSE), then the confidence intervals are added to the plot}

\item{plot_risk}{if TRUE (default), then the risk table is added to be plot}
}
\value{

}
\description{
Generate Kaplan-Meier Plot
}
\examples{

setDbOptions(getSettings())

data <- getPatientAnnotation()
ca <- exampleClassAssigment(high = 0.16)
ca  <- classAssignment2df(ca)
dataSurv <- inner_join(data, ca)

surv <- survivalAnalysis(dataSurv)
generateKaplanMeierPlot(surv)
generateKaplanMeierPlot(surv, plot_ci = TRUE)

}
