% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateMetabolicsPlot}
\alias{generateMetabolicsPlot}
\title{Generate Metabolics Plot}
\usage{
generateMetabolicsPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame created from \code{\link{getTissueDataMetabolicsById}}}

\item{plotType}{one of the follwoing "point", "roc", "violin", "box", "coverage"
or "all" to get a list with all plots}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataMetabolicsById, then this
parameter does not need to be specified.}
}
\value{
ggplot
}
\description{
Generate Metabolics Plot
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.16)
data <- getTissueDataMetabolicsById("gluconeogenesis", ca)

generateMetabolicsPlot(data, "all")

}
