% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateMetabolicsWaterfallPlot}
\alias{generateMetabolicsWaterfallPlot}
\title{Generate Metabolics Waterfall Plot}
\usage{
generateMetabolicsWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame created from \code{\link{getWaterfallDataMetabolics}}}

\item{fill}{column from data used to fill value}
}
\value{
ggplot
}
\description{
Generate Metabolics Waterfall Plot
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma"))
data <- getWaterfallDataMetabolics("gluconeogenesis", prefilter)

generateMetabolicsWaterfallPlot(data)

}
