% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateMutationalBurdenWaterfallPlot}
\alias{generateMutationalBurdenWaterfallPlot}
\title{Generate waterfall plot for mutational burden}
\usage{
generateMutationalBurdenWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame, output of getWaterfallDataMutationalBurden}

\item{fill}{character string, variable to use for bars coloring}
}
\value{
ggplot object
}
\description{
Produces a waterfall plot with mutational burden data
}
\examples{
setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
df <- getWaterfallDataMutationalBurden(prefilter)
generateMutationalBurdenWaterfallPlot(df)

}
