% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateMutationalFractionPlot}
\alias{generateMutationalFractionPlot}
\title{Generate mutational fraction plot by type}
\usage{
generateMutationalFractionPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame, output of getTissueDataMutationalBurden()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataMutationalBurden, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces mutational fraction plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage" and "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{
setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.5)
anno <- getTissueAnnotation()
df <- getTissueDataMutationalBurden(ca, anno)
label <- "Mutational fraction"
generateMutationalFractionPlot(df, "point", paste(label, "point plot"))
generateMutationalFractionPlot(df, "roc", paste(label, "roc plot"))
generateMutationalFractionPlot(df, "violin", paste(label, "violin plot"))
generateMutationalFractionPlot(df, "box", paste(label, "box plot"))
generateMutationalFractionPlot(df, "coverage", paste(label, "data coverage"))

}
