% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateOtherPropertiesPlot}
\alias{generateOtherPropertiesPlot}
\title{Generate Other Properties Plot}
\usage{
generateOtherPropertiesPlot(sampleClasses, sAnno, property, classLabel = NULL)
}
\arguments{
\item{sampleClasses}{classAssigment object}

\item{sAnno}{tissue annotation (see: \code{getTissueAnnotation})}

\item{property}{which property to be plotted}

\item{classLabel}{if sampleClasses is a list instead of classAssigment
then the classLabels needs to be provided.}
}
\value{
ggplot object
}
\description{
Generate Other Properties Plot
}
\examples{
setDbOptions(getSettings())
sampleClasses <- exampleClassAssigment()
sAnno <- getTissueAnnotation()
generateOtherPropertiesPlot(sampleClasses, sAnno, "lossofy")
generateOtherPropertiesPlot(sampleClasses, sAnno, "organ")

}
