% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateProteinExpressionPlot}
\alias{generateProteinExpressionPlot}
\title{Generate protein expression plot by type}
\usage{
generateProteinExpressionPlot(df, plotType, title = NULL, ca = NULL)
}
\arguments{
\item{df}{data.frame, output of getTissueDataRppaById()}

\item{plotType}{character string, plot type}

\item{title}{character string, the plot title}

\item{ca}{classAssignment object if it cannot be inferred from \code{df} param.
However if df is an output of getTissueDataRppaById, then this
parameter does not need to be specified.}
}
\value{
ggplot object
}
\description{
Produces protein expression plot that shows class comparison. Available types: 
"point" (default), "roc", "violin", "box", "coverage", "all". Coverage plot may
be used to show the data availability in each class.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(0.15,0.2)
antibody <- "MEK1_pS217_S221"
df <- getTissueDataRppaById(antibody, ca)
generateProteinExpressionPlot(df, "point", paste(antibody, "point plot"))
generateProteinExpressionPlot(df, "roc", paste(antibody, "roc plot"))
generateProteinExpressionPlot(df, "violin", paste(antibody, "violin plot"))
generateProteinExpressionPlot(df, "box",  paste(antibody, "box plot"))
generateProteinExpressionPlot(df, "coverage", paste(antibody, "data coverage"))
generateProteinExpressionPlot(df, "all")

}
