% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateProteinWaterfallPlot}
\alias{generateProteinWaterfallPlot}
\title{Generate Protein Waterfall Plot}
\usage{
generateProteinWaterfallPlot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame created by \code{\link{getWaterfallDataRppa}}}

\item{fill}{column name from data to be used for filling the bars}
}
\value{
ggplot2 object
}
\description{
Generate Protein Waterfall Plot
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
data <- getWaterfallDataRppa("CDK1", prefilter)
generateProteinWaterfallPlot(data)

}
