% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{generateSignalingInputBarplot}
\alias{generateSignalingInputBarplot}
\title{Generate Signaling Input Barplot}
\usage{
generateSignalingInputBarplot(data, fill = "tumortype")
}
\arguments{
\item{data}{data.frame created with \code{\link{getInputDataSignaling}}}

\item{fill}{column from data used to fill value}
}
\value{
ggplot2
}
\description{
Generate Signaling Input Barplot
}
\examples{

setDbOptions(getSettings())

selected <- c("cholangiocarcinoma", "lymphoid neoplasm diffuse large b-cell lymphoma", 
  "uterine carcinosarcoma", "kidney chromophobe", "adrenocortical carcinoma", 
  "uveal melanoma")

prefilter <- makePrefilter(selected = "uveal melanoma")
pathways <- getAvailableSignalingPathways()
data <- getInputDataSignaling(pathways[1], prefilter)
generateSignalingInputBarplot(data)

}
