% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getAnalysisDataMutation}
\alias{getAnalysisDataMutation}
\title{Get Analysis Mutation Data}
\usage{
getAnalysisDataMutation(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissues or classAssigment object}
}
\value{
data.frame with columns: ensg, tissuename
}
\description{
Get Analysis Mutation Data
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
df <- getAnalysisDataMutation(ca)
head(df)

}
