% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getAvailableGeneSignatures}
\alias{getAvailableGeneSignatures}
\title{Get Available Gene Signatures}
\usage{
getAvailableGeneSignatures()
}
\value{
data.frame with columns: signature, description, hyperlink
}
\description{
Get Available Gene Signatures
}
\examples{

setDbOptions(getSettings())
getAvailableGeneSignatures()

}
