% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getAvailableHallmarkGeneSets}
\alias{getAvailableHallmarkGeneSets}
\title{Get Available Hallmark Gene Sets}
\usage{
getAvailableHallmarkGeneSets()
}
\value{
vector of available gene sets
}
\description{
Get Available Hallmark Gene Sets
}
\examples{

setDbOptions(getSettings())
getAvailableHallmarkGeneSets()

}
