% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getAvailableImmuneCellTypes}
\alias{getAvailableImmuneCellTypes}
\title{Get Available Immune Cell Types}
\usage{
getAvailableImmuneCellTypes()
}
\value{

}
\description{
Get Available Immune Cell Types
}
\examples{

setDbOptions(getSettings())
getAvailableImmuneCellTypes()

}
