% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getAvailableMetabolicPathways}
\alias{getAvailableMetabolicPathways}
\title{Get Available Metabolic Pathways}
\usage{
getAvailableMetabolicPathways()
}
\value{
vector with all available metabolic pathways
}
\description{
Get Available Metabolic Pathways
}
\examples{

setDbOptions(getSettings())
getAvailableMetabolicPathways()

}
