% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getExpressionDimRedData}
\alias{getExpressionDimRedData}
\title{Get Data for Dimension Reduction of Gene Expression}
\usage{
getExpressionDimRedData(sampleClasses, ensg_gene_set = NA, p = FALSE)
}
\arguments{
\item{sampleClasses}{classAssignment object, or list with two classes.}

\item{ensg_gene_set}{NA (default, returns the most varying genes) or
vector of gene ensgs}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
list with elements:
- data.counts - matrix with gene expression counts
- data.log2tpm - matrix with gene expression log2tpm
- assignment - data.frame with tissuename and class assignment
}
\description{
Get Data for Dimension Reduction of Gene Expression
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
res <- getExpressionDimRedData(ca, p = TRUE)
str(res)

resGeneSet <- getExpressionDimRedData(ca, "HALLMARK_P53_PATHWAY", p = TRUE)
str(resGeneSet)

}
