% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getExpressionDimRedDataNGenes}
\alias{getExpressionDimRedDataNGenes}
\title{Get Gene Expression Data in a Long Format for Most Varying Genes}
\usage{
getExpressionDimRedDataNGenes(tissuenames, limit = 1000)
}
\arguments{
\item{tissuenames}{classAssignment object or vector of tissue names}

\item{limit}{maximum number of most varying genes to be returned}
}
\value{
data.frame with columns: tissuenames, ensg, counts,
log2tpm
}
\description{
Get Gene Expression Data in a Long Format for Most Varying Genes
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
head(getExpressionDimRedDataNGenes(ca, 10))

}
