% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getGeneAnno}
\alias{getGeneAnno}
\title{Get gene annotation}
\usage{
getGeneAnno(species)
}
\arguments{
\item{species}{string containing the species name}
}
\value{
list of data.frames
}
\description{
Get all available gene annotation for the selected species.
The function returns as a data.frame with fields:
ensg, symbol, name, location
}
\examples{

setDbOptions(getSettings())
head(getGeneAnno("human"))

}
