% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getGeneFromSymbol}
\alias{getGeneFromSymbol}
\title{Get Gene information from Symbol}
\usage{
getGeneFromSymbol(mySymbol, species)
}
\arguments{
\item{mySymbol}{gene symbol}

\item{species}{species name}
}
\value{
list with values: ensg, name, symbol
}
\description{
Get Gene information from Symbol
}
\examples{

setDbOptions(getSettings())
getGeneFromSymbol("KRAS", "human")

}
