% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{getGSEA}
\alias{getGSEA}
\title{Gene Set Enrichment Analysis}
\usage{
getGSEA(diffExResult, geneSets, rankType, p = FALSE, minVal = 1e-300)
}
\arguments{
\item{diffExResult}{data.frame, table result obtained from 
differentialGeneExpression_LimmaVoom() function}

\item{geneSets}{list of gene IDs, gene sets}

\item{rankType}{character string, "p.valueDir" or "logFC"}
}
\value{
tibble with columns: p.geomean, stat.mean, p.val, q.val, set.size, 
exp1, geneset
}
\description{
Calculate GSEA for differential gene expression results
}
\examples{
\donttest{
setDbOptions(getSettings())
tissueClasses <- exampleClassAssigment()
anno <- getGeneAnno("human")

dge <- differentialGeneExpression_LimmaVoom(tissueClasses, anno)
geneSets <- getGSEAdata("human", "hallmark")
getGSEA(dge, geneSets, "p.valueDir", p = TRUE)
}
}
