% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getHeatmapDataHallmark}
\alias{getHeatmapDataHallmark}
\title{Get Data for Hallmark Genesets Heatmap}
\usage{
getHeatmapDataHallmark(tissueClasses, score = c("gsva", "ssgsea"))
}
\arguments{
\item{tissueClasses}{classAssignment object or list with two tissuename classes}

\item{score}{character string, score to retrieve from DB. Available choices:
gsva (default), ssgsea}
}
\value{
matrix with gene sets in rows and tissuename in columns
}
\description{
Get Data for Hallmark Genesets Heatmap
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
getHeatmapDataHallmark(ca)[1:5,1:3]

}
