% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getInfoMutation}
\alias{getInfoMutation}
\title{Get mutation information}
\usage{
getInfoMutation(ensg, prefilter)
}
\arguments{
\item{ensg}{character string, gene ID}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with fields: tissuename, tumortype, dnamutation, aamutation,
aamutated, histology_type, histology_subtype
}
\description{
Returns mutation information related to the specified gene ID
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
ensg <- "ENSG00000133703"
getInfoMutation(ensg, prefilter)

}
