% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getInformation}
\alias{getInformation}
\title{Get Data Base Information}
\usage{
getInformation()
}
\value{
data.frame with columns: Data Source, Version
}
\description{
Get Data Base Information
}
\examples{

setDbOptions(getSettings())
getInformation()

}
