% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getInputDataSignaling}
\alias{getInputDataSignaling}
\title{Get Input Data Signaling}
\usage{
getInputDataSignaling(pathway, prefilter)
}
\arguments{
\item{pathway}{string containing pathway 
(see \code{\link{getAvailableSignalingPathways}} for all available options)}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with tissuename, {pathway}, and tumortype.
}
\description{
Get Input Data Signaling
}
\examples{

setDbOptions(getSettings())
pathways <- getAvailableSignalingPathways()
prefilter <- makePrefilter("PDX", selected = "adrenocortical carcinoma")
head(getInputDataSignaling(pathways[1], prefilter))

}
