% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{getMultipleFeaturesAxisData}
\alias{getMultipleFeaturesAxisData}
\title{Get Multiple Features Axis Data}
\usage{
getMultipleFeaturesAxisData(sourceType, gene, axis, prefilter)
}
\arguments{
\item{sourceType}{data source type ('expr' or 'copy')}

\item{gene}{ensg or list with gene ensg and symbol}

\item{axis}{name of the target axis ('x' or 'y')}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
list with varLabel and scale name, and data.frame with
tissuename {x/y}Var and tumortype.
}
\description{
Get Multiple Features Axis Data
}
\examples{

setDbOptions(getSettings())

prefilter <- makePrefilter()
getMultipleFeaturesAxisData("expr", "ENSG00000133703", "x", prefilter)

}
