% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{getMutationAssociation}
\alias{getMutationAssociation}
\title{Get mutation association}
\usage{
getMutationAssociation(tissueClasses, anno, p = FALSE)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{anno}{data.frame, gene annotation}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
tibble with columns: ensg, symbol, name, class2, class1, higher, 
p.value, adj.P.Value
}
\description{
Compares mutation data for 2 tissue classes.
}
\examples{
setDbOptions(getSettings())
tissueClasses <- exampleClassAssigment(high = 0.3)
anno <- getGeneAnno("human")

getMutationAssociation(tissueClasses, anno, p = TRUE)

}
