% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getOtherPropData}
\alias{getOtherPropData}
\title{Add other proportions data to tissue annotation}
\usage{
getOtherPropData(
  tissuename,
  tissueAnno,
  p = FALSE,
  mutationalFractionThresholds = c(0.1, 0.5)
)
}
\arguments{
\item{tissuename}{character vector with tissue names}

\item{tissueAnno}{tissue annotation data.frame}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}

\item{mutationalFractionThresholds}{thresholds used to determine cutoffs for
\code{mutational_fraction_in_percent} column}
}
\value{
data.frame with the all the columns from \code{anno} and new:

- rnasequenced,
- copynumbered,
- mutational_fraction_in_percent,
- mutational_burden_class
- tumor_purity_class
}
\description{
Add other proportions data to tissue annotation
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
anno <- getTissueAnnotation()
head(getOtherPropData(ca, anno))

}
