% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{getOtherPropStatistics}
\alias{getOtherPropStatistics}
\title{Get other properties statistics}
\usage{
getOtherPropStatistics(sampleClasses, sAnno, p = FALSE)
}
\arguments{
\item{sampleClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{sAnno}{data.frame, tissue annotation}

\item{p}{logical, ShinySession or FutureManager task object; used for progress}
}
\value{
data.frame with columns: property, p.value, method, estimate, data, 
statistic, higher
}
\description{
Calculates statistics for continuous variables (Wilcoxon test) and class 
variables (Fisher's exact test) available in annotation.
}
\examples{
setDbOptions(getSettings())
tissueClasses <- exampleClassAssigment(low = 0.05, tt = NULL)
anno <- getTissueAnnotation()

getOtherPropStatistics(tissueClasses, anno, p = TRUE)

}
