% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_utils.R
\name{getPatientAnnotationFilter}
\alias{getPatientAnnotationFilter}
\title{Filter Patient Annotation Data by Prefilter}
\usage{
getPatientAnnotationFilter(prefilter, pa)
}
\arguments{
\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}

\item{pa}{patient annotation data (see \code{\link{getPatientAnnotation}})}
}
\value{
filtered patient annotation data by applying tissue panel and
selected entries from TiffPrefilter object.
}
\description{
Filter Patient Annotation Data by Prefilter
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("PDX", selected = "adrenocortical carcinoma")
pa <- getPatientAnnotation()
getPatientAnnotationFilter(prefilter, pa)

}
