% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getSequencedTissues}
\alias{getSequencedTissues}
\title{Get Sequenced Tissues}
\usage{
getSequencedTissues(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissuenames or classAssigment object}
}
\value{

}
\description{
Get Sequenced Tissues
}
\examples{

setDbOptions(getSettings())
head(getSequencedTissues(exampleClassAssigment()))

}
