% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{getSettings}
\alias{getSettings}
\title{Get TIFF settings}
\usage{
getSettings(obj = NULL)
}
\arguments{
\item{obj}{NULL or the settings object}
}
\description{
Retrieves settings from an yml file. By default, the file from the package 
is used. If you want to use your custom file, please set TIFF_CONFIG_FILE 
environmental variable.
}
\details{
If the `obj` argument is not NULL, it will be returned from the function.
}
