% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueAnnotation}
\alias{getTissueAnnotation}
\title{Get Tissue Annotation}
\usage{
getTissueAnnotation()
}
\value{
data.frame with tissue annotation.
}
\description{
Get Tissue Annotation
}
\examples{

setDbOptions(getSettings())
head(getTissueAnnotation())

}
