% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{getTissueCounts}
\alias{getTissueCounts}
\title{Get tissue expression data}
\usage{
getTissueCounts(tissueClasses)
}
\arguments{
\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}
}
\value{
list with fields: expr, tpm, and design
}
\description{
This is a helper function that retrieves expression data for the specified 
tissue.
}
\examples{

setDbOptions(getSettings())
tissueClasses <- exampleClassAssigment()
tissueCounts <- getTissueCounts(tissueClasses)

}
