% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataClones}
\alias{getTissueDataClones}
\title{Get Tissue Data Clones}
\usage{
getTissueDataClones(tissuenames, scores = "clone_tree_score")
}
\arguments{
\item{tissuenames}{vector of tissuenames or classAssigment object}

\item{scores}{vector of scores to be extracted}
}
\value{
data.frame
}
\description{
Get Tissue Data Clones
}
\examples{

setDbOptions(getSettings())

getTissueDataClones(NULL)

}
