% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataCopyNumber}
\alias{getTissueDataCopyNumber}
\title{Get Tissue Copy Number Data}
\usage{
getTissueDataCopyNumber(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissuenames or classAssigment}
}
\value{
data.frame with columns: tissuenames, ensg and log2relativecopynumber
}
\description{
Get Tissue Copy Number Data
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
head(getTissueDataCopyNumber(ca))

}
