% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneExpression}
\alias{getTissueDataGeneExpression}
\title{Get Tissue Gene Expression Data}
\usage{
getTissueDataGeneExpression(tissuenames, unit = "log2tpm")
}
\arguments{
\item{tissuenames}{character vector with tissuenames}

\item{unit}{unit of the result, possible values are `log2tpm` (default)
or `counts`.}
}
\value{
data.frame with columns: tissuenames, ensg, log2tpm
}
\description{
Get Tissue Gene Expression Data
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
head(getTissueDataGeneExpression(ca))
head(getTissueDataGeneExpression(ca, "counts"))

}
