% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneExpressionById}
\alias{getTissueDataGeneExpressionById}
\title{Get tissue Gene Expression data by gene ID}
\usage{
getTissueDataGeneExpressionById(
  ensg,
  tissueClasses,
  addRownames = TRUE,
  unit = c("tpm", "log2tpm", "counts")
)
}
\arguments{
\item{ensg}{character string, gene ID. Use NULL to get data for all genes}

\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{addRownames}{if TRUE (default), the tissue names are added as rownames
to the output table}

\item{unit}{unit of the result, possible values are `tpm` (default)
or `log2tpm`.}
}
\value{
data.frame with fields: tissuename, ensg, tpm
}
\description{
Returns Gene Expression data for the specified tissue and gene.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"
getTissueDataGeneExpressionById(ensg, ca)

}
