% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneSet}
\alias{getTissueDataGeneSet}
\title{Get tissue Gene Set data}
\usage{
getTissueDataGeneSet(tissuenames, score = c("gsva", "ssgsea"))
}
\arguments{
\item{tissuenames}{character vector of tissue names}

\item{score}{character string, score to retrieve from DB. Available choices:
gsva (default), ssgsea}
}
\value{
data.frame with columns:tissuename, gene_set, score
}
\description{
SQL wrapper for Gene Set data retrieval.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
getTissueDataGeneSet(ca)

}
