% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneSetById}
\alias{getTissueDataGeneSetById}
\title{Get tissue Gene Set data by set ID}
\usage{
getTissueDataGeneSetById(geneSet, tissueClasses, score = c("gsva", "ssgsea"))
}
\arguments{
\item{geneSet}{character string, set ID. Use NULL to get data for all sets}

\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{score}{character string, score to retrieve from DB. Available choices:
gsva (default), ssgsea}
}
\value{
data.frame with fields: tissuename, gene_set, !!score
}
\description{
Returns Gene Set data for the specified tissue and set ID
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
getTissueDataGeneSetById("HALLMARK_P53_PATHWAY", ca)

}
