% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneSignature}
\alias{getTissueDataGeneSignature}
\title{Get tissue Gene Signature data}
\usage{
getTissueDataGeneSignature(tissuenames)
}
\arguments{
\item{tissuenames}{character vector of tissue names}
}
\value{
data.frame with columns: tissuename, signature, score
}
\description{
SQL wrapper for Gene Signature data retrieval.
}
\examples{
setDbOptions(getSettings())
ca <- exampleClassAssigment()
getTissueDataGeneSignature(ca)

}
