% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataGeneSignatureById}
\alias{getTissueDataGeneSignatureById}
\title{Get tissue Gene Signature data by signature ID}
\usage{
getTissueDataGeneSignatureById(signature, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{signature}{character string, signature ID. Use NULL to get data for all 
signatures}

\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{addRownames}{if TRUE (default), the tissue names are added as rownames
to the output table}
}
\value{
data.frame with fields: tissuename , signature, score
}
\description{
Returns Gene Signature data for the specified tissue and signature ID
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
geneSig <- getTissueDataGeneSignatureById(NULL, ca)

}
