% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataImmuneCells}
\alias{getTissueDataImmuneCells}
\title{Get Tissue Data Immune Cells}
\usage{
getTissueDataImmuneCells(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissuenames or classAssigment object}
}
\value{
data.frame with columns tissuenames, celltype and score.
}
\description{
Get Tissue Data Immune Cells
}
\examples{

setDbOptions(getSettings())
head(getTissueDataImmuneCells(exampleClassAssigment()))

}
