% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataImmuneCellsById}
\alias{getTissueDataImmuneCellsById}
\title{Get Tissue Data Immune Cells By Id}
\usage{
getTissueDataImmuneCellsById(cellType, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{cellType}{\code{\link{getAvailableImmuneCellTypes()}}}

\item{tissueClasses}{vector of tissuenames or classAssigment object}

\item{addRownames}{if TRUE (default) tissuenames are added 
as rownames to the output table.}
}
\value{
data.frame with columns tissuename, celltype and score.
}
\description{
Get Tissue Data Immune Cells By Id
}
\examples{

setDbOptions(getSettings())
head(getTissueDataImmuneCellsById("Neurons", NULL))
head(getTissueDataImmuneCellsById("Neurons", exampleClassAssigment()))

}
