% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataMetabolics}
\alias{getTissueDataMetabolics}
\title{Get Tissue Data Metabolics}
\usage{
getTissueDataMetabolics(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissuenames or classAssigment object}
}
\value{
data.frame with tissuename, metabolic_pathway and score.
}
\description{
Get Tissue Data Metabolics
}
\examples{

setDbOptions(getSettings())
head(getTissueDataMetabolics(exampleClassAssigment()))

}
