% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataMetabolicsById}
\alias{getTissueDataMetabolicsById}
\title{Get Tissue Data Metabolics By Id}
\usage{
getTissueDataMetabolicsById(pathway, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{pathway}{metabolic pathway (see \code{\link{getAvailableMetabolicPathways}})}

\item{tissueClasses}{classAssignment object or vector with tissue names. If NULL
returns all tissues}

\item{addRownames}{if TRUE (default) tissuenames are added 
as rownames to the output table.}
}
\value{

}
\description{
Get Tissue Data Metabolics By Id
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter()
head(getTissueDataMetabolicsById("gluconeogenesis", NULL))

getTissueDataMetabolicsById("gluconeogenesis", c("TCGA-GU-A42R-01", "TCGA-KQ-A41Q-01"))

}
