% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{getTissueDataMutationalBurden}
\alias{getTissueDataMutationalBurden}
\title{Get Tissue Data Mutational Burden for generateMutationalFractionPlot}
\usage{
getTissueDataMutationalBurden(ca, anno)
}
\arguments{
\item{ca}{tissue class assigment}

\item{anno}{tissue annotation}
}
\value{
data.frame with `tissuename` and `value` columns
}
\description{
Get Tissue Data Mutational Burden for generateMutationalFractionPlot
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment(high = 0.5)
anno <- getTissueAnnotation()
df <- getTissueDataMutationalBurden(ca, anno)

}
