% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataMutationById}
\alias{getTissueDataMutationById}
\title{Get tissue Mutation data by gene ID}
\usage{
getTissueDataMutationById(ensg, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{ensg}{character string, gene ID. Use NULL to get data for all genes}

\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{addRownames}{if TRUE then adds tissuename as rownames}
}
\value{
data.frame with fields: tissuename, ensg, aamutation, aamutated, 
rnazygosity, dnazygosity
}
\description{
Returns Mutation data for the specified tissue and gene.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"

getTissueDataMutationById(ensg, ca)

}
