% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataRppa}
\alias{getTissueDataRppa}
\title{Get tissue RPPA data}
\usage{
getTissueDataRppa(tissuenames)
}
\arguments{
\item{tissuenames}{character vector of tissue names or classAssigment object}
}
\value{
data.frame with columns: tissuename, antibody, score
}
\description{
SQL wrapper for RPPA data retrieval.
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
getTissueDataRppa(ca)

}
