% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataRppaById}
\alias{getTissueDataRppaById}
\title{Get tissue RPPA data by antibody ID}
\usage{
getTissueDataRppaById(antibody, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{antibody}{character string, antibody ID. Use NULL to get data for all 
antibodies}

\item{tissueClasses}{classAssignment object or a 
list with fields: class1 and class2. Each field should
contain a character vector of tissue names}

\item{addRownames}{if TRUE (default), the tissue names are added as rownames
to the output table}
}
\value{
data.frame with fields: tissuename, antibody, score
}
\description{
Returns RPPA data for the specified tissue and antibody
}
\examples{

setDbOptions(getSettings())
ca <- exampleClassAssigment()
ensg <- "ENSG00000133703"
getTissueDataRppaById(NULL, ca)

}
