% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataSignaling}
\alias{getTissueDataSignaling}
\title{Get Tissue Data Signaling}
\usage{
getTissueDataSignaling(tissuenames)
}
\arguments{
\item{tissuenames}{vector of tissue names}
}
\value{
data.frame with columns tissuename, pathway, active.
}
\description{
Get Tissue Data Signaling
}
\examples{

setDbOptions(getSettings())
getTissueDataSignaling(c("TCGA-06-2557-01", "TCGA-06-2569-01"))

}
