% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueDataSignalingById}
\alias{getTissueDataSignalingById}
\title{Get Tissue Data Signaling By Id}
\usage{
getTissueDataSignalingById(pathway, tissueClasses, addRownames = TRUE)
}
\arguments{
\item{pathway}{signaling pathway (see \code{\link{getAvailableSignalingPathways}})}

\item{tissueClasses}{vector with selected tissue names, if NULL returns all
available ones}

\item{addRownames}{if TRUE (default), the tissue names are added as rownames
to the output table}
}
\value{

}
\description{
Get Tissue Data Signaling By Id
}
\examples{

setDbOptions(getSettings())
head(getTissueDataSignalingById("cell_cycle", NULL))

}
