% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getTissueGroups}
\alias{getTissueGroups}
\title{Get Tissue Groups}
\usage{
getTissueGroups()
}
\value{
list with tissue groups
}
\description{
Get Tissue Groups
}
\examples{

setDbOptions(getSettings())
getTissueGroups()

}
