% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_utils.R
\name{getTissuenameTranslation}
\alias{getTissuenameTranslation}
\title{Get tissue name translation and add the tissue annotation}
\usage{
getTissuenameTranslation(ti, ti_anno)
}
\arguments{
\item{ti}{data.frame with tissuename column}

\item{ti_anno}{data.frame, tissue annotation}
}
\value{
list with fields: df, missing
}
\description{
Get tissue name translation and add the tissue annotation
}
\details{
This is an utility function for merging a data.frame (\code{ti})
with tissue annotation data.frame (\code{ti_anno}). It returns a list
with merged data.frame and and character vector of tissue names that are
not available in annotation data.
}
\examples{

setDbOptions(getSettings())

ti_anno <- getTissueAnnotation()
ti <- exampleClassAssigment()
ti$class2 <- c(ti$class2, "MISSING") # add non-existing tissuename
ti <- XIFF::stackClasses(ti)

translated <- getTissuenameTranslation(ti, ti_anno)
translated$missing

}
