% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataCopyNumber}
\alias{getWaterfallDataCopyNumber}
\title{Get Waterfall Data Copy Number}
\usage{
getWaterfallDataCopyNumber(ensg, prefilter)
}
\arguments{
\item{ensg}{gene ensg}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with tissuename, relativecopynumber and tumortype
}
\description{
Get Waterfall Data Copy Number
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter()
data <- getWaterfallDataCopyNumber("ENSG00000133703", prefilter)
head(data)

}
