% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataGeneSignatures}
\alias{getWaterfallDataGeneSignatures}
\title{Get Waterfall Data Gene Signatures}
\usage{
getWaterfallDataGeneSignatures(signature, prefilter)
}
\arguments{
\item{signature}{gene signatures (see \code{\link{getAvailableGeneSignatures}})}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with tissuename, score and tumortype
}
\description{
Get Waterfall Data Gene Signatures
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter()
head(getWaterfallDataGeneSignatures("HRD", prefilter))

}
