% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataHallmarkSet}
\alias{getWaterfallDataHallmarkSet}
\title{Get waterfall Hallmark Set data}
\usage{
getWaterfallDataHallmarkSet(geneSet, prefilter, score = c("gsva", "ssgsea"))
}
\arguments{
\item{geneSet}{character string, gene set ID}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}

\item{score}{character string, score to retrieve from DB. Available choices:
gsva (default), ssgsea}
}
\value{
data.frame with columns: tissuename, tumortype, score
}
\description{
SQL wrapper for Hallmark Set data retrieval. The `tissuename` column 
in the result table is a factor with levels sorted according to the 
`score` value (desc order).
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
getWaterfallDataHallmarkSet("HALLMARK_P53_PATHWAY", prefilter)

}
