% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataImmuneCells}
\alias{getWaterfallDataImmuneCells}
\title{Get Waterfall Data Immune Cells}
\usage{
getWaterfallDataImmuneCells(cellType, prefilter)
}
\arguments{
\item{cellType}{string with cell type}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with columns 'tissuename', 'score' and tumortype
}
\description{
Get Waterfall Data Immune Cells
}
\examples{

setDbOptions(getSettings())
df <- getWaterfallDataImmuneCells("Activated dendritic cells", makePrefilter())
head(df)

}
