% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataMetabolics}
\alias{getWaterfallDataMetabolics}
\title{Get Waterfall Data for Metabolics}
\usage{
getWaterfallDataMetabolics(pathway, prefilter)
}
\arguments{
\item{pathway}{metabolic pathway (see \code{\link{getAvailableMetabolicPathways}})}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with tissuename, score and tumortype.
}
\description{
Get Waterfall Data for Metabolics
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter()
head(getWaterfallDataMetabolics("gluconeogenesis", prefilter))

}
