% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataMutationalBurden}
\alias{getWaterfallDataMutationalBurden}
\title{Get waterfall Mutational Burden data}
\usage{
getWaterfallDataMutationalBurden(prefilter)
}
\arguments{
\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with columns: tissuename, tumortype, mutational_fraction
}
\description{
SQL wrapper for Mutational Burden data retrieval. The `tissuename` column 
in the result table is a factor with levels sorted according to the 
`mutational_fraction` value (desc order).
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
getWaterfallDataMutationalBurden(prefilter)

}
