% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{getWaterfallDataRppa}
\alias{getWaterfallDataRppa}
\title{Get waterfall Protein RPPA data}
\usage{
getWaterfallDataRppa(antibody, prefilter)
}
\arguments{
\item{antibody}{character string, antibody ID}

\item{prefilter}{TiffPrefilter object (see \code{\link{makePrefilter}})}
}
\value{
data.frame with columns: tissuename, tumortype, x_score
}
\description{
SQL wrapper for Protein RPPA data retrieval. The `tissuename` column 
in the result table is a factor with levels sorted according to the 
`x_score value` (desc order).
}
\examples{

setDbOptions(getSettings())
prefilter <- makePrefilter("tumor", c("thymoma", "mesothelioma"))
getWaterfallDataRppa("CDK1", prefilter)

}
