% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_plots.R
\name{grid.draw.ggsurvplot}
\alias{grid.draw.ggsurvplot}
\title{Grid draw method for ggsurvplot}
\usage{
\method{grid.draw}{ggsurvplot}(x)
}
\arguments{
\item{x}{ggsurvplot object}
}
\value{
nothing, called for side effects.
}
\description{
Grid draw method for ggsurvplot
}
\details{
required to properly draw ggsurvplot in shiny application.
}
