% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{initialSettingsCheck}
\alias{initialSettingsCheck}
\title{Check the integrity of the settings object}
\usage{
initialSettingsCheck(settings)
}
\arguments{
\item{settings}{settings object}
}
\value{
TRUE if settings object is valid.
}
\description{
Check the integrity of the settings object
}
\examples{

config <- getSettings()
initialSettingsCheck(config)

}
