% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{makePrefilter}
\alias{makePrefilter}
\title{Makes prefilter object}
\usage{
makePrefilter(
  tissueType = c("tumor", "adjacent normal", "normal", "PDX"),
  selected = NULL
)
}
\arguments{
\item{tissueType}{string denoting tissue type}

\item{selected}{vector of selected tissue types}
}
\value{
TiffPrefilter object
}
\description{
Makes prefilter object
}
\examples{

makePrefilter("tumor")

}
