% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{makeTextCallback}
\alias{makeTextCallback}
\title{Utility function for making the text callback function.}
\usage{
makeTextCallback(scoreTemplate, sigx = 4, sigy = 4)
}
\arguments{
\item{scoreTemplate}{glue template which will render the text.}

\item{sigx}{indicating the number of decimal places for rx.}

\item{sigy}{indicating the number of decimal places for ry.}
}
\value{

}
\description{
Utility function for making the text callback function.
}
\examples{

cl <- makeTextCallback( "Cxpression score from {ry[2]} to {ry[1]}")
cl(100, c(0,10), c(0, 100))
cl(100, c(0,10), c(0, NA))

}
