% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_db.R
\name{print.TiffPrefilter}
\alias{print.TiffPrefilter}
\title{Print TiffPrefilter object}
\usage{
\method{print}{TiffPrefilter}(x, width = NULL, ...)
}
\arguments{
\item{x}{TiffPrefilter object}

\item{width}{max number of characters in a line. Used for printing. Default
value should be good in most cases.}

\item{...}{other arguments passed to the print method, discarded.}
}
\value{

}
\description{
Print TiffPrefilter object
}
\examples{
x <- makePrefilter("tumor")
x

}
