% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%||\%}
\alias{setDbOptions}
\alias{getPostgresql}
\alias{classAssignment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{XIFF}{\code{\link[XIFF:reexports]{\%>\%}}, \code{\link[XIFF:op-null-default]{\%||\%}}, \code{\link[XIFF]{classAssignment}}, \code{\link[XIFF]{getPostgresql}}, \code{\link[XIFF]{setDbOptions}}}
}}

