% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the TIFF application}
\usage{
run(...)
}
\arguments{
\item{...}{additional params passed into \code{shiny::runApp function}.}
}
\value{
nothing, called for side effects
}
\description{
Run the TIFF application
}
