% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo_analysis.R
\name{survivalAnalysis}
\alias{survivalAnalysis}
\title{Perform Survival Analysis}
\usage{
survivalAnalysis(data)
}
\arguments{
\item{data}{data.frame with days_to_death, days_to_last_followup and class 
columns.}
}
\value{
list with \code{df} object used to fit the model and \code{fit}
containing survfit object.
}
\description{
Perform Survival Analysis
}
\examples{

setDbOptions(getSettings())

data <- getPatientAnnotation()
ca <- exampleClassAssigment()
ca  <- classAssignment2df(ca)
dataSurv <- inner_join(data, ca)
survivalAnalysis(dataSurv)

}
